refactor: move to the deprecation dumpster
