+++ frappe/email/doctype/email_queue/email_queue.py
@@ -32,7 +32,6 @@ from frappe.utils import (
 	sbool,
 	split_emails,
 )
-from frappe.utils.deprecations import deprecated
 from frappe.utils.verified_command import get_signed_params
 
 
@@ -223,15 +222,9 @@ class EmailQueue(Document):
 		).run()
 
 
-@task(queue="short")
-@deprecated
-def send_mail(email_queue_name, smtp_server_instance: SMTPServer = None):
-	"""This is equivalent to EmailQueue.send.
+from frappe.deprecation_dumpster import send_mail as _send_mail
 
-	This provides a way to make sending mail as a background job.
-	"""
-	record = EmailQueue.find(email_queue_name)
-	record.send(smtp_server_instance=smtp_server_instance)
+send_mail = task(queue="short")(_send_mail)
 
 
 class SendMailContext:
