+++ frappe/geo/country_info.py
@@ -8,7 +8,6 @@ import os
 from functools import lru_cache
 
 import frappe
-from frappe.utils.deprecations import deprecated
 from frappe.utils.momentjs import get_all_timezones
 
 
@@ -39,9 +38,7 @@ def _get_country_timezone_info():
 	return {"country_info": get_all(), "all_timezones": get_all_timezones()}
 
 
-@deprecated
-def get_translated_dict():
-	return get_translated_countries()
+from frappe.deprecation_dumpster import get_translated_dict
 
 
 def get_translated_countries():
