+++ frappe/utils/__init__.py
@@ -23,9 +23,10 @@ from typing import TypedDict
 
 from werkzeug.test import Client
 
+from frappe.deprecation_dumpster import gzip_compress, gzip_decompress, make_esc
+
 # utility functions like cint, int, flt, etc.
 from frappe.utils.data import *
-from frappe.utils.deprecations import deprecated
 from frappe.utils.html_utils import sanitize_html
 
 EMAIL_NAME_PATTERN = re.compile(r"[^A-Za-z0-9\u00C0-\u024F\/\_\' ]+")
@@ -421,14 +422,6 @@ def get_file_timestamp(fn):
 			raise
 		else:
 			return None
-
-
-# to be deprecated
-def make_esc(esc_chars):
-	"""
-	Function generator for Escaping special characters
-	"""
-	return lambda s: "".join("\\" + c if c in esc_chars else c for c in s)
 
 
 # esc / unescape characters -- used for command line
@@ -886,34 +879,6 @@ def call(fn, *args, **kwargs):
 	                bench --site erpnext.local execute frappe.utils.call --args '''["frappe.get_all", "Activity Log"]''' --kwargs '''{"fields": ["user", "creation", "full_name"], "filters":{"Operation": "Login", "Status": "Success"}, "limit": "10"}'''
 	"""
 	return json.loads(frappe.as_json(frappe.call(fn, *args, **kwargs)))
-
-
-# Following methods are aken as-is from Python 3 codebase
-# since gzip.compress and gzip.decompress are not available in Python 2.7
-
-
-@deprecated
-def gzip_compress(data, compresslevel=9):
-	"""Compress data in one shot and return the compressed string.
-	Optional argument is the compression level, in range of 0-9.
-	"""
-	from gzip import GzipFile
-
-	buf = io.BytesIO()
-	with GzipFile(fileobj=buf, mode="wb", compresslevel=compresslevel) as f:
-		f.write(data)
-	return buf.getvalue()
-
-
-@deprecated
-def gzip_decompress(data):
-	"""Decompress a gzip compressed string in one shot.
-	Return the decompressed string.
-	"""
-	from gzip import GzipFile
-
-	with GzipFile(fileobj=io.BytesIO(data)) as f:
-		return f.read()
 
 
 def get_safe_filters(filters):
